analogs.test {downscaleR}	R Documentation

Donwscaling with the analogs method (test).

Description

Donwscaling with the analogs method (test).

Usage

analogs.test(newdata, x, y, dates, info)

Arguments

newdata
The grid data. Class: matrix.

x
The grid data where to search the analogs. Class: matrix.

y
The observations data of the same days of x. Class: matrix.

dates
A list containing both the newdata grid and x grid, dates.

info
A list containing the information required to perform the analogs method: number of analogs, function, window. These parameters are grouped in a list in analogs.train.

Details

The selected functions use the base R functions: mean, median, weighted.mean, quantile, Extremes.

Value

A matrix containing the predictions.

Author(s)

J. Bano-Medina

[Package downscaleR version 3.0.1 Index]