biasCorrection1D {downscaleR}	R Documentation

Bias correction methods on 1D data

Description

Implementation of several standard bias correction methods

Usage

biasCorrection1D(o, p, s, method, scaling.type, fitdistr.args, precip,
  pr.threshold, n.quantiles, extrapolation, theta, parallel = FALSE,
  max.ncores = 16, ncores = NULL)

Arguments

o
A vector (e.g. station data) containing the observed climate data for the training period

p
A vector containing the simulated climate by the model for the training period.

s
A vector containing the simulated climate for the variable used in p, but considering the test period.

method
method applied. Current accepted values are "eqm", "delta", "scaling", "pqm", "gpqm", "variance", "loci" and "ptr".

scaling.type
Character indicating the type of the scaling method. Options are "additive" (default) or "multiplicative" (see details). This argument is ignored if "scaling" is not selected as the bias correction method.

fitdistr.args
Distribution to use when applying the "pqm" method (parametric quantile mapping). Choices are "gamma" or "gaussian" (see details).

precip
Logical indicating if o, p, s is precipitation data.

pr.threshold
The minimum value that is considered as a non-zero precipitation. Ignored for varcode values different from "pr". Default to 1 (assuming mm).

extrapolation
Character indicating the extrapolation method to be applied to correct values in "s" that are out of the range of "p". Extrapolation is applied only to the "eqm" method, thus, this argument is ignored if other bias correction method is selected. Default is "none" (do not extrapolate).

theta
numeric indicating upper threshold (and lower for the left tail of the distributions, if needed) above which precipitation (temperature) values are fitted to a Generalized Pareto Distribution (GPD). Values below this threshold are fitted to a gamma (normal) distribution. By default, 'theta' is the 95th percentile (5th percentile for the left tail). Only for "gpqm" method.

parallel
Logical. Should parallel execution be used?

max.ncores
Integer. Upper bound for user-defined number of cores.

ncores
Integer number of cores used in parallel computation. Self-selected number of cores is used when ncpus = NULL (the default), or when maxcores exceeds the default ncores value.

Author(s)

M. Iturbide

[Package downscaleR version 3.0.1 Index]