delta {downscaleR}	R Documentation

Delta method for bias correction

Description

Implementation of Delta method for bias correction

Usage

delta(o, p, s)

Arguments

o
A vector (e.g. station data) containing the observed climate data for the training period

p
A vector containing the simulated climate by the model for the training period.

s
A vector containing the simulated climate for the variable used in x, but considering the test period.

Author(s)

S. Herrera and M. Iturbide

[Package downscaleR version 3.0.1 Index]