downs.predict {downscaleR}	R Documentation

Switch to selected downscale method.

Description

Internal function of downscale.predict that switches to the corresponding method.

Usage

downs.predict(x, method, atomic_model)

Arguments

x
The grid data. Class: matrix.

method
The method of the given model.

atomic_model
An object containing the statistical model of the selected method.

Details

This function is internal and should not be used by the user. The user should use downscale.predict.

Value

A matrix with the predictions.

Author(s)

J. Bano-Medina

[Package downscaleR version 3.0.1 Index]