downs.train {downscaleR}	R Documentation

Switch to selected downscale method.

Description

Internal function of downscale.train that switches to the selected method.

Usage

downs.train(x, y, method, ...)

Arguments

x
The input grid. Class: matrix.

y
The observations dataset. Class: matrix.

method
Type of transer function. Options are: analogs, GLM and NN.

...
Optional parameters. These parameters are different depending on the method selected. Every parameter has a default value set in the atomic functions in case that no selection is wanted. For this reason see the atomic functions for more details: glm.train and nn.train.

Details

The optional parameters of neural networks can be found in the library deepnet via nn.trainThis function is internal and should not be used by the user. The user should use downscale.train.

Value

An object with the information of the selected model.

Author(s)

J. Bano-Medina

[Package downscaleR version 3.0.1 Index]