downscale.cv {downscaleR}	R Documentation

Downscale climate data and reconstruct the temporal serie by splitting the data following a user-defined scheme

Description

Downscale climate data and reconstruct the temporal serie by splitting the data following a user-defined scheme. The statistical downscaling methods currently implemented are: analogs, generalized linear models (GLM) and Neural Networks (NN).

Usage

downscale.cv(x, y, method, sampling.strategy = "kfold.chronological",
  folds = 4, scale.list = NULL, global.vars = NULL,
  combined.only = TRUE, spatial.predictors = NULL,
  local.predictors = NULL, extended.predictors = NULL, condition = NULL,
  threshold = NULL, ...)

Arguments

x
The input grid (admits both single and multigrid, see makeMultiGrid). It should be an object as returned by loadeR.

y
The observations dataset. It should be an object as returned by loadeR.

method
A string value. Type of transer function. Currently implemented options are "analogs", "GLM" and "NN".

sampling.strategy
Specifies a sampling strategy to define the training and test subsets. Possible values are "kfold.chronological" (the default), "kfold.random" and "leave-one-year-out". The sampling.strategy choices are next described:

"kfold.random" creates the number of folds indicated in the folds argument by randomly sampling the entries along the time dimension.

"kfold.chronological" is similar to "kfold.random", but the sampling is performed in ascending order along the time dimension.

"leave-one-year-out". This schema performs a leave-one-year-out cross validation. It is equivalent to introduce in the argument folds a list of all years one by one.

The first two choices will be controlled by the argument folds (see below)

folds
This arguments controls the number of folds, or how these folds are created (ignored if sampling.strategy = "leave-one-year-out"). If it is given as a fraction in the range (0-1), it splits the data in two subsets, one for training and one for testing, being the given value the fraction of the data used for training (i.e., 0.75 will split the data so that 75% of the instances are used for training, and the remaining 25% for testing). In case it is an integer value (the default, which is 4), it sets the number of folds in which the data will be split (e.g., folds = 10 for the classical 10-fold cross validation). Alternatively, this argument can be passed as a list, each element of the list being a vector of years to be included in each fold (See examples).

scale.list
A list of the parameters related to scale grids. This parameter calls the function scaleGrid. See the function definition for details on the parameters accepted.

global.vars
An optional character vector with the short names of the variables of the input x multigrid to be retained as global predictors (use the getVarNames helper if not sure about variable names). This argument just produces a call to subsetGrid, but it is included here for better flexibility in downscaling experiments (predictor screening...). For instance, it allows to use some specific variables contained in x as local predictors and the remaining ones, specified in subset.vars, as either raw global predictors or to construct the combined PC.

combined.only
Optional, and only used if spatial.predictors parameters are passed. Should the combined PC be used as the only global predictor? Default to TRUE. Otherwise, the combined PC constructed with which.combine argument in prinComp is append to the PCs of the remaining variables within the grid.

spatial.predictors
Default to NULL, and not used. Otherwise, a named list of arguments in the form argument = value, with the arguments to be passed to prinComp to perform Principal Component Analysis of the predictors grid (x). See Details on principal component analysis of predictors.

local.predictors
Default to NULL, and not used. Otherwise, a named list of arguments in the form argument = value, with the following arguments:

vars: names of the variables in x to be used as local predictors

fun: Optional. Aggregation function for the selected local neighbours. The aggregation function is specified as a list, indicating the name of the aggregation function in first place (as character), and other optional arguments to be passed to the aggregation function. For instance, to compute the average skipping missing values: fun = list(FUN= "mean", na.rm = TRUE). Default to NULL, meaning that no aggregation is performed.

n: Integer. Number of nearest neighbours to use. If a single value is introduced, and there is more than one variable in vars, the same value is used for all variables. Otherwise, this should be a vector of the same length as vars to indicate a different number of nearest neighbours for different variables.

extended.predictors
This is a parameter related to the extreme learning machine and reservoir computing framework where input data is randomly projected into a new space of size n. Default to NULL, and not used. Otherwise, a named list of arguments in the form argument = value, with the following arguments:

n: A numeric value. Indicates the size of the random nonlinear dimension where the input data is projected.

module: A numeric value (Optional). Indicates the size of the mask's module. Belongs to a specific type of ELM called RF-ELM.

condition
Inequality operator to be applied considering the given threshold. "GT" = greater than the value of threshold, "GE" = greater or equal, "LT" = lower than, "LE" = lower or equal than. We only train with the days that satisfy the condition.

threshold
Numeric value. Threshold used as reference for the condition. Default is NULL. If a threshold value is supplied with no specification of the parameter condition. Then condition is set to "GE".

...
Optional parameters. These parameters are different depending on the method selected. Every parameter has a default value set in the atomic functions in case that no selection is wanted. Everything concerning these parameters is explained in the section Details of the function downscale.train. However, if wanted, the atomic functions can be seen here: glm.train and nn.train.

Details

The function relies on prepareData, prepareNewData, downscale.train, and downscale.predict. For more information please visit these functions. It is envisaged to allow for a flexible fine-tuning of the cross-validation scheme. It uses internally the transformeR helper dataSplit for flexible data folding. Note that the indices for data splitting are obtained using getYearsAsINDEX when needed (e.g. in leave-one-year-out cross validation), thus adequately handling potential inconsistencies in year selection when dealing with year-crossing seasons (e.g. DJF).

If the variable to downscale is the precipitation and it is a binary variable, then two temporal series will be returned:

The temporal serie with binary values filtered by a threshold adjusted by the train dataset, see binaryGrid for more details.

The temporal serie with the results obtained by the downscaling, without binary conversion process.

Missing data removal is recommended prior to multisite calibration.

According to the concept of cross-validation, a particular year should not appear in more than one fold (i.e., folds should constitute disjoint sets). For example, the choice fold =list(c(1988,1989), c(1989, 1990)) will raise an error, as 1989 appears in more than one fold.

Value

The reconstructed downscaled temporal serie.

Author(s)

J. Bano-Medina

See Also

downscale.train for training a downscaling model downscale.predict for prediction for a a test dataset with a trained model for downscaleR Wiki for downscaling seasonal forecasting and climate projections.

Other downscaling.functions: downscale.predict, downscale.train, downscale

Examples

require(transformeR)
data(NCEP_Iberia_hus850, NCEP_Iberia_ta850)
x <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_ta850)
x <- subsetGrid(x, years = 1985:1995)
# Loading predictands
y <- VALUE_Iberia_pr
y <- getTemporalIntersection(obs = y, prd = x, "obs")
x <- getTemporalIntersection(obs = y, prd = x, "prd")
# ... kfold in 3 parts equally divided ...
pred <- downscale.cv(x, y, folds = 3, sampling.strategy = "kfold.chronological",
                     scale.list = list(type = "standardize"),
                     method = "GLM", family = Gamma(link = "log"), condition = "GT", threshold = 0,
                     spatial.predictors = list(which.combine = getVarNames(x), v.exp = 0.9))
# ... kfold by years ...
pred <- downscale.cv(x,y,sampling.strategy = "kfold.chronological",
                     method = "GLM", condition = "GT", threshold = 0,
                     scale.list = list(type = "standardize"),
                     folds = list(c(1985,1986,1987,1988),
                                  c(1989,1990,1991,1992),
                                  c(1993,1994,1995)))
# ... leave one year out  ...
pred <- downscale.cv(x,y,sampling.strategy = "leave-one-year-out",
                     method = "GLM", condition = "GT", threshold = 0,
                     scale.list = list(type = "standardize"))
# Reconstructing the downscaled serie in 3 folds with local predictors.
pred <- downscale.cv(x,y,folds = 3, sampling.strategy = "kfold.chronological",
                     scale.list = list(type = "standardize"),
                     method = "GLM", condition = "GT", threshold = 0,
                     local.predictors = list(vars = "hus@850", n = 4))

[Package downscaleR version 3.0.1 Index]