flatMemberDim {downscaleR}	R Documentation

Concatenate members

Description

Concatenate members as a single time series for using their joint distribution in bias correction

Usage

flatMemberDim(grid)

Arguments

grid
Input (multimember) grid

Value

A grid without members, with additional attributes to retrieve the original structure after bias correction

Author(s)

J Bedia

See Also

recoverMemberDim, for recovering the original structure after bias correction.

[Package downscaleR version 3.0.1 Index]