getWindowIndex {downscaleR}	R Documentation

Get the index of the window days.

Description

Get the index of the days that corresponding to the window and the target days centered in it.

Usage

getWindowIndex(y, x, newdata, window, delta.method = FALSE)

Arguments

y
A grid or station data containing the observed climate data for the training period

newdata
A grid containing the simulated climate for the test period.

window
vector of length = 2 (or 1) specifying the time window width used to calibrate and the target days (days that are being corrected). If the window length = 1 the window width is no larger than the target days. The window is centered on the target day/s (window width >= target days).

delta.method
logical (default is FALSE)

method
method applied. Current accepted values are "eqm", "delta", "scaling", "pqm" and "gpqm" "variance","loci" and "ptr". See details.

Author(s)

M. Iturbide

[Package downscaleR version 3.0.1 Index]