glimpr {downscaleR}	R Documentation

GLM downscaling

Description

Implementation of GLM's to downscale precipitation data

Usage

glimpr(y = y, modelPars = modelPars, simulate = c("none", "loocv",
  "kfold"), return.models = FALSE, wet.threshold = wet.threshold,
  n.pcs = n.pcs)

Arguments

y
A grid or station data containing the observed climate data for the training period

modelPars
Output object from function ppModelSetup containing the predictors and test data

simulate
Character. Options are "no", "yes" or "occurrence". The last option simulates the occurrence but not the amount.

wet.threshold
Value below which precipitation amount is considered zero

n.eofs
Integer indicating the number of EOFs to be used as predictors

Author(s)

J Bedia, M Iturbide

See Also

Other downscaling: biasCorrection, isimip

[Package downscaleR version 3.0.1 Index]