glm.predict {downscaleR}	R Documentation

Donwscaling with a given generalized linear model (GLM).

Description

Donwscaling with a given generalized linear models (GLM) calculated in downscale.predict via glm.train.

Usage

glm.predict(x, weights, info)

Arguments

x
The grid data. Class: matrix.

weights
Object as returned from glm.train

info
A list containing information of the experiment: the fitting, the family of the generalized linear model and if it is deterministic or stochastic.

Details

Predicts by using the base function predict. This function is internal and should not be used by the user. The user should use downscale.predict.

Value

The predicted matrix.

Author(s)

J. Bano-Medina

[Package downscaleR version 3.0.1 Index]