gpqm {downscaleR}	R Documentation

Generalized Quantile Mapping method for bias correction

Description

Implementation of Generalized Quantile Mapping method for bias correction

Usage

gpqm(o, p, s, precip, pr.threshold, theta)

Arguments

o
A vector (e.g. station data) containing the observed climate data for the training period

p
A vector containing the simulated climate by the model for the training period.

s
A vector containing the simulated climate for the variable used in p, but considering the test period.

precip
Logical indicating if o, p, s is precipitation data.

pr.threshold
The minimum value that is considered as a non-zero precipitation. Ignored for varcode values different from "pr". Default to 1 (assuming mm).

theta
numeric indicating upper threshold (and lower for the left tail of the distributions, if needed) above which precipitation (temperature) values are fitted to a Generalized Pareto Distribution (GPD). Values below this threshold are fitted to a gamma (normal) distribution. By default, 'theta' is the 95th percentile (5th percentile for the left tail). Only for "gpqm" method.

Author(s)

S. Herrera and M. Iturbide

[Package downscaleR version 3.0.1 Index]