loci {downscaleR}	R Documentation

Local intensity scaling of precipitation

Description

Implementation of Local intensity scaling of precipitation method for bias correction based on Vincent Moron's local_scaling function in weaclim toolbox in Matlab

Usage

loci(o, p, s, precip, pr.threshold)

Arguments

o
A vector (e.g. station data) containing the observed climate data for the training period

p
A vector containing the simulated climate by the model for the training or test period.

s
A vector containing the simulated climate for the variable used in p, but considering the test period.

precip
Logical indicating if o, p, s is precipitation data.

pr.threshold
The minimum value that is considered as a non-zero precipitation.

Author(s)

B. Szabo-Takacs

[Package downscaleR version 3.0.1 Index]