norain {downscaleR}	R Documentation

norain

Description

presprocess to bias correct precipitation data

Usage

norain(o, p, threshold)

Arguments

o
A vector (e.g. station data) containing the observed climate data for the training period

p
A vector containing the simulated climate by the model for the training period.

threshold
The minimum value that is considered as a non-zero precipitation. Ignored for varcode values different from "pr". Default to 1 (assuming mm).

Author(s)

S. Herrera and M. Iturbide

[Package downscaleR version 3.0.1 Index]