pqm {downscaleR}	R Documentation

Parametric Quantile Mapping method for bias correction

Description

Implementation of Parametric Quantile Mapping method for bias correction

Usage

pqm(o, p, s, fitdistr.args, precip, pr.threshold)

Arguments

o
A vector (e.g. station data) containing the observed climate data for the training period

p
A vector containing the simulated climate by the model for the training period.

s
A vector containing the simulated climate for the variable used in x, but considering the test period.

fitdistr.args
Further arguments passed to function fitdistr (densfun, start, ...). Only used when applying the "pqm" method (parametric quantile mapping). Please, read the fitdistr help document carefully before parameter setting in fitdistr.args.

precip
Logical indicating if o, p, s is precipitation data.

pr.threshold
The minimum value that is considered as a non-zero precipitation. Ignored for varcode values different from "pr". Default to 1 (assuming mm).

Author(s)

S. Herrera and M. Iturbide

[Package downscaleR version 3.0.1 Index]