predictor.nn.values {downscaleR}	R Documentation

Construct local predictor matrices

Description

Constructs the local predictor matrices given their spatial index position (as returned by predictor.nn.indices) and additional parameters.

Usage

predictor.nn.values(nn.indices.list, grid, fun = NULL)

Arguments

nn.indices.list
A list of index positions as returned by predictor.nn.indices

grid
The grid containing the local predictor information. It can be either the predictors (training) grid or the simulation (prediction) grid, depending on the situation.

fun
A list of funtions for neighbour aggregation for each local predictor variable. Default to NULL, and no aggregation is performed

Value

A nested list of 2D matrices with the following structure: sites/members

Author(s)

J Bedia

See Also

Other downscaling.helpers: predictor.nn.indices, prepareData, prepareNewData

[Package downscaleR version 3.0.1 Index]