prepareNewData {downscaleR}	R Documentation

Prepare newdata for predictions

Description

Prepare the prediction data according to the definition of the experiment

Usage

prepareNewData(newdata, data.structure)

Arguments

newdata
A grid containing the prediction data.

data.structure
A structure, as returned by prepareData

Value

A named list with the components required by the downscaling method in order to perform the predictions

Author(s)

J Bedia

See Also

downscaleR Wiki for preparing predictors for downscaling and seasonal forecasting.

Other downscaling.helpers: predictor.nn.indices, predictor.nn.values, prepareData

Examples

# Loading data
data("VALUE_Iberia_tas")
y <- VALUE_Iberia_tas 
data("NCEP_Iberia_hus850", "NCEP_Iberia_psl", "NCEP_Iberia_ta850")
x <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_psl, NCEP_Iberia_ta850)
# Calculating EOFs
data <- prepareData(x = x, y = y, spatial.predictors = list(v.exp = 0.95))
# Projecting a new dataset to the calculated EOFs
newdata <- prepareNewData(x,data)

[Package downscaleR version 3.0.1 Index]