quickDiagnostics {downscaleR}	R Documentation

Plots daily/annual series and the annual correlation map of different grid objects

Description

Plots daily/annual series and the annual correlation map of different grid objects

Usage

quickDiagnostics(obs, raw, downscaled = NULL, location = NULL,
  type = c("daily", "interannual"), members = NULL, na.tolerance = 0.3,
  ylim = NULL, main = NULL)

Arguments

obs
grid of observations.

raw
grid of model data.

downscaled
Optional. grid of the downscaling output.

location
Coordinates of a location in the geographic domain of 'obs'. If NULL it is randomly chosen from coordinates in 'obs'.

type
Character value, either "daily" or "interannual", indicating is the assessment is to be performed on a daily or interannual basis.

members
An integer vector indicating the position of the members to be subset. If NULL all members are considered when type = "interannual" and one member is randomly chosen when type = "daily".

na.tolerance
proportion of NAs in a grid cell (location) that are allowed to calculate correlation.

ylim
'ylim' argument passed to the time series plot.

main
'main' argument passed to the plot.

Value

Two diagnostic plots with observed, simulated and (possibly) downscaled time series, and a QQ-plot by percentlies.

Author(s)

M. Iturbide

Examples

{
data("VALUE_Iberia_pr")
data("NCEP_Iberia_pr")
y <- VALUE_Iberia_pr
x <- NCEP_Iberia_pr
x$Data <- x$Data*86400
quickDiagnostics(obs = y, raw = x, location = c(-2, 43), type = "daily")
#quickDiagnostics(obs = y, raw = x, location = c(-2, 43), type = "interannual")
eqm1win <- biasCorrection(y = y, x = x,
                          method = "eqm",
                          extrapolation = "none")
quickDiagnostics(y, x, eqm1win)
quickDiagnostics(y, x, eqm1win, location = c(-2, 43))
#quickDiagnostics(obs = y, raw = x, downscaled = eqm1win, 
# location = c(-2, 43), type = "interannual")
}

[Package downscaleR version 3.0.1 Index]