scaling {downscaleR}	R Documentation

Scaling method for bias correction

Description

Implementation of Scaling method for bias correction

Usage

scaling(o, p, s, scaling.type)

Arguments

o
A vector (e.g. station data) containing the observed climate data for the training period

p
A vector containing the simulated climate by the model for the training period.

s
A vector containing the simulated climate for the variable used in x, but considering the test period.

scaling.type
Character indicating the type of the scaling method. Options are "additive" (default) or "multiplicative" (see details). This argument is ignored if "scaling" is not selected as the bias correction method.

Author(s)

S. Herrera and M. Iturbide

[Package downscaleR version 3.0.1 Index]