variance {downscaleR}	R Documentation

Variance scaling of temperature

Description

Implementation of Variance scaling of temperature method for bias correction

Usage

variance(o, p, s, precip)

Arguments

o
A vector (e.g. station data) containing the observed climate data for the training period

p
A vector containing the simulated climate by the model for the training period.

s
A vector containing the simulated climate for the variable used in p, but considering the test period.

precip
Logical indicating if o, p, s is temperature data.

Author(s)

B. Szabo-Takacs

[Package downscaleR version 3.0.1 Index]